move neighbor to bond
